refactored controllers for resume and portfolio combined both controllers into one single controller
