--- src/main/java/com/codeup/controllers/Resume.java
@@ -1,15 +1,0 @@
-package com.codeup.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-
-@Controller
-public class Resume {
-
-    @GetMapping("/resume")
-    public String welcome() {
-        return "profile/resume";
-    }
-}
